requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-dotenv
aiosmtplib
pandas
streamlit
email-validator